+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - Issue #55: Add configure proxy scan interval
+- Issue #69: Add physic measures
 
 ### Fixed